+++ tests/CBenchmark.h
@@ -1,4 +1,4 @@
 void bytewrite(unsigned char *a, int bytes);
 unsigned char byteread(unsigned char *a, int bytes);
-void wordwrite(unsigned int *a, int bytes);
-unsigned int wordread(unsigned int *a, int bytes);
+void wordwrite(unsigned long *a, int bytes);
+unsigned int wordread(unsigned long *a, int bytes);
